fix(ancs): do not let a TLV length wrap the response offset
response_is_complete walked the Data Source TLVs with a uint16_t
offset. A length field of 0xFFF8 or more (the phone is the GATT server
and every byte it sends is untrusted) wraps off back below resp_len,
the check passes, the TLV is counted, and a truncated response is
declared complete. ancs_finish_command uses int arithmetic so it fails
safe and Lua sees an empty attribute table, but the real fragments
then arrive in IDLE and are dropped, or get glued onto the next
command and run it into the 15 s timeout.

Make the offset 32-bit; no other change.

+++ modules/halo/src/ble_ancs.c
@@ -790,7 +790,10 @@ static void ancs_finish_command(int status)
  * Called with ctx.lock held. Returns true when complete. */
 static bool response_is_complete(void)
 {
-	uint16_t off;
+	/* 32-bit on purpose: a hostile or corrupt TLV length near 0xFFFF would
+	 * wrap a uint16_t offset back below resp_len, pass the check in the
+	 * loop below and declare a truncated response complete. */
+	uint32_t off;
 	uint8_t tlv_count = 0;
 	uint8_t expected;
